Implement annotation destroy authorization
References BiodataMiningGroup/dias-core#5

+++ app/Http/Controllers/Api/AnnotationController.php
@@ -99,7 +99,7 @@ class AnnotationController extends Controller
     public function destroy($id)
     {
         $annotation = Annotation::findOrFail($id);
-        $this->requireCanEdit($annotation);
+        $this->authorize('destroy', $annotation);
 
         $annotation->delete();
